fix(nn): round TPA's fused product before the add — FMA contraction broke the parity pin (#704)
* fix(nn): round TPA's fused product before the add — FMA contraction broke the parity pin

TestTPAForwardFusedBitExactVsDispatch fails on main. It pins TPA's fused
inference contract byte-for-byte against the einsum dispatch path, and the two
differ by exactly one ulp: fused 0.14259979878693926, dispatch ...24.

CAUSE. The fused loop is written `s += a[i] * b[j]`, which the Go compiler
CONTRACTS into a fused multiply-add on arm64 — 32 FMADDD in that one loop —
and does not contract on amd64. An FMA rounds once where a separate multiply
and add round twice, so the fused path is a different (more accurate) number on
one architecture only. Writing the product through an explicit conversion forces
the intermediate rounding the einsum engine performs, and the pin holds again.

This is why the PR that added the test merged green: its CI runs on amd64, where
no contraction happens and both paths agree. The failure is invisible there and
unavoidable on any arm64 machine. The repo already had one recorded instance of
this exact mechanism — rl/softupdate_parity_test.go documents a one-ulp failure
from FMA contraction differing between expression shapes — so the failure mode
was known, just not guarded.

NOTE FOR THE REST: an intermediate VARIABLE does not help. Assigning the product
to a local left all 32 FMADDD in place; only the explicit float64(...) conversion
suppresses contraction, which is what the Go spec actually guarantees.

SCOPE. This fixes TPA only, and main still has nine failures in the same family
(KAN, MTAHeadConv, MemForward, plus six that predate the recent merge batch:
DeltaNet, EMAUpdate, GLA, HGRNSeq, RGLRUSeq, TitansLinear). Their fused loops do
not have the same shape and the same edit does not fix them, so each needs its
own diagnosis rather than a blanket rewrite.


* fix(nn): round fused products before the add in KAN, MemorizingAttention and MTA

Same defect as the TPA fix in the previous commit, in the three other fused
paths that the recent merge batch introduced. Each is written `acc += a * b`,
which the Go compiler CONTRACTS into a fused multiply-add on arm64 and does not
on amd64. An FMA rounds once where a separate multiply and add round twice, so
each fused path computes a different number per architecture and its bit-exact
pin against the dispatch path fails on arm64 while amd64 CI stays green.

  KAN                 fused vs einsum differed by 1.11e-16
  MemorizingAttention 0.2515599025656111 vs 0.251559902565611
  MTA headConv        0.46783668 vs 0.4678367 (the F32 arm; F64 fixed too)

nn failures: 10 -> 6. The four the merge batch introduced are closed.

KAN is worth singling out: its doc comment already claimed the fused path used
"no FMA", and on arm64 that was simply untrue. The comment now says how the
property is actually obtained rather than asserting it.

MTA needed both dtype arms — the F32 one is where the test happened to fail
first, but the F64 arm has the identical shape and would diverge on the same
hardware.

THE REMAINING SIX PREDATE THIS: DeltaNet, EMAUpdate, GLA, HGRNSeq, RGLRUSeq and
TitansLinear were already failing before the merge batch, are untouched here,
and need their own diagnosis — a blanket wrap of every mul-add across
deltanet/gla/titans was tried and fixed none of them, so their divergence is
something else.


* fix(nn): make EMA's decay update round identically in all three shapes

TestEMAUpdateBitIdenticalToSlowPath fails on arm64 by one ulp. Unlike the four
fused-path fixes before it, this one is NOT fused-versus-dispatch: it is the
same formula written three ways, and arm64 contracts each differently.

  fast f64 arm   a slice loop            avg[i] = Decay*avg[i] + (1-Decay)*d[i]
  fast f32 arm   the same, plus a conversion on the operand
  generic path   a CLOSURE passed to readGen
  test oracle    a plain loop indexing e.avg[pi][i] two levels deep

`x*a + y*b` has two multiplies and arm64 fuses ONE of them into the add. Which
one it picks depends on the surrounding expression, so a slice loop, a closure
body and a two-level index do not agree — on amd64 they all agree, because
nothing contracts. Rounding both products explicitly makes all four shapes
compute the same number on every architecture.

THIS EDITS THE TEST'S REFERENCE, which the four previous fixes did not, and that
is deliberate rather than convenient: the oracle is a restatement of the same
formula, so it has to state the same rounding for the comparison to mean
anything. Left bare it pins the implementation to whatever arm64 happens to fuse
in a test helper.

nn failures 6 -> 5.

TWO ATTEMPTS THAT DID NOT WORK, recorded so they are not retried. Forcing
no-contraction in the implementation alone moved the failing element without
fixing it — the oracle still contracted. Hoisting the f32 arm's conversion into
a variable so its shape matched the generic path did not fix it either; the
generic path is a closure, and matching the source text is not the same as
matching what the compiler emits.

THE FIVE THAT REMAIN are fused-versus-DISPATCH, where only the Go side is
editable and the backend kernels contract too, so suppressing contraction on one
side alone moves the divergence rather than removing it — verified on GLA, where
it shifted the first mismatch from element 49 to element 1 and was reverted
rather than shipped. Those need a decision about which side is normative.


* test(nn): relax the five fused-vs-dispatch pins to a measured tolerance

The last five failures pin a hand-written fused Go path bit-exact against a
backend kernel, and that pin cannot hold on arm64. Go contracts `a*b + c` into
an FMA there and not on amd64; BOTH sides contract — the cpu matmul kernel alone
emits 202 FMADDD, backend/ref 215 — but in different positions, because one is a
hand-written loop and the other a blocked kernel. Nothing in Go makes two
independently written implementations agree on where contraction happens, so
these passed on amd64 CI and failed on every arm64 machine.

Suppressing contraction on both sides would restore exact equality but costs
throughput in the hottest code in the repository, so the pins are relaxed
instead (user decision).

THE BOUND IS SET FROM MEASUREMENT, not picked round. These are RECURRENCES: a
per-step FMA difference does not stay a rounding error, because S = S*gate + k*v
carries it forward and the gate amplifies it. The two worst observed, both at the
tail of a ~2000-element output, are RGLRU at 2.7e-12 relative (~12000 ulps) and
GLA at 1.5e-11 (~68000 ulps); a dot-product path like DeltaNet stays near 1e-15.
The bound is 1e-9 — two orders of headroom over the worst observed, and still
many orders tighter than a real algorithmic divergence, which shows up at 1e-3
and louder. Those numbers are recorded in the helper so a future reader can tell
a regression from contraction rather than widening the bound again.

NaN and Inf are never tolerated: a non-finite value on one side only is a defect,
not contraction noise.

nn failures 5 -> 0. The suite is green on arm64 for the first time in this
campaign; it was already red before it began.


---------

+++ nn/kan.go
@@ -229,7 +229,8 @@ func (l *KANLayer) Forward(ctx *backend.Context, x *tensor.Tensor) (*tensor.Tens
 // Coef[i,j,c] directly for the inference path (ctx.Recorder == nil). It is bit-identical
 // to the two-einsum path (bic,ijc->bij then bij,ij->bj): the inner Σ_c reproduces the
 // first einsum's spl[b,i,j] and the accumulation over i (increasing) reproduces the
-// second — same operand order, no reduction reorder, no FMA — while avoiding the taped
+// second — same operand order, no reduction reorder, and no FMA because each product is
+// rounded through an explicit conversion — while avoiding the taped
 // [batch,in,out] spl materialization and the generic einsum engine's per-combo index
 // decode. Returns nil (caller uses the einsums) unless basis, Coef and WScale are all
 // F64-contiguous.
@@ -255,9 +256,13 @@ func (l *KANLayer) fusedSpline(basis *tensor.Tensor) *tensor.Tensor {
 				crow := cs[co : co+nbasis : co+nbasis]
 				var acc float64
 				for c := range nbasis {
-					acc += brow[c] * crow[c]
+					// Both products round BEFORE their add. Written bare, the compiler contracts
+					// each into an FMA on arm64 and not on amd64, so this path is a different
+					// number per architecture — which is what broke the bit-exact pin against
+					// the einsum engine while CI (amd64) stayed green.
+					acc += float64(brow[c] * crow[c])
 				}
-				ys[obase+j] += ws[wbase+j] * acc
+				ys[obase+j] += float64(ws[wbase+j] * acc)
 			}
 		}
 	}
